backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
opencv-python
numpy
requests

# LLM Summaries (Phase 3)
google-genai

# JWT Auth (Phase 4)
PyJWT